claw-agent/requirements.txt
playwright==1.43.0
pytest-playwright==0.5.0
openai==1.14.3
anthropic==0.21.3
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
pydantic==2.6.4
google-generativeai==0.5.4
